Initial update - basic blocks
Most of this is a piecemeal vision of what I am trying to do. Some files are entirely made up of comments and there is no folder structure. As I gain more understanding of Dash and Plotly, the structure of this project may change so drastically that I might need to restart with a proper set up, but for now I will charge forward with my ideas.